# server/swagger.yaml
openapi: 3.0.3
info:
  title: Volleyball Scoreboard API
  description: |
    API для управления табло волейбольного матча и системой авторизации.

    ## Провайдеры данных
    Система поддерживает два провайдера, переключаемых через переменную окружения `DB_PROVIDER`:
    - **firebase** — Firebase Firestore + Firebase Auth
    - **pocketbase** — PocketBase (self-hosted)

    ## Аутентификация
    1. Получите токен через `POST /api/auth/login`
    2. Используйте его в заголовке `Authorization: Bearer <token>` для защищённых эндпоинтов
    Все эндпоинты `/api/scoreboard/*` и `/api/matches/*` требуют авторизации.
    Только `/api/auth/login` и `/health` доступны без токена.

    ## Управление пользователями
    Создание, редактирование и удаление пользователей осуществляется **не через API**,
    а через админ-панель провайдера:
    - **Firebase Console** → Authentication → Users
    - **PocketBase Admin** → `http://your-server:8090/_/` → Collections → scoreusers
  version: 1.0.0
  contact:
    name: Volleyball Scoreboard
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
...
security:
  - BearerAuth: []
paths:
...
  /api/scoreboard:
    post:
      summary: Создать новую игру
      description: |
        Создаёт новое табло с указанным `game_id`.
        Возвращает 409 Conflict если игра с таким ID уже существует.
        Все поля кроме `game_id` опциональны — используются значения по умолчанию.
      tags:
        - Scoreboard
      operationId: createScoreboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScoreboard'
      responses:
        '201':
          description: Игра создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scoreboard'
        '400':
          description: Отсутствует game_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Игра с таким ID уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/scoreboard/{game_id}:
    get:
      summary: Получить состояние табло
      description: Возвращает все поля текущего матча
      tags:
        - Scoreboard
      operationId: getScoreboard
      parameters:
        - name: game_id
          in: path
          required: true
          schema:
            type: string
          description: Идентификатор матча
      responses:
        '200':
          description: Состояние табло
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scoreboard'
        '404':
          description: Матч не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Обновить произвольные поля табло
      description: Позволяет обновить любые поля, кроме `lastEdited` (устанавливается автоматически)
      tags:
        - Scoreboard
      operationId: updateScoreboard
      parameters:
        - name: game_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoreboardUpdate'
      responses:
        '200':
          description: Обновлённое состояние табло
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scoreboard'
  /api/scoreboard/{game_id}/score:
    post:
      summary: Изменить счёт (добавить/убрать очко)
      description: |
        Основное действие — изменение счёта команды. Автоматически обрабатывает:
        - Проверку победы в сете
        - Автоматическую смену площадок (пляжный — каждые N суммарных очков, классический — 8 очков на тай-брейке)
        - Фиксацию истории сетов
        - Сохранение результата матча при завершении
      tags:
        - Scoreboard
      operationId: updateScore
      parameters:
        - name: game_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - team
                - delta
              properties:
                team:
                  type: string
                  enum: [home, away]
                  description: Команда, которой начисляется/снимается очко
                delta:
                  type: integer
                  enum: [1, -1]
                  description: +1 — добавить очко, -1 — убрать
      responses:
        '200':
          description: Обновлённое состояние табло
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scoreboard'
        '400':
          description: Ошибка валидации (отрицательный счёт и т.д.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/scoreboard/{game_id}/new-set:
    post:
      summary: Начать новый сет
      description: Обнуляет счёт, инкрементирует номер сета
      tags:
        - Scoreboard
      operationId: newSet
      parameters:
        - name: game_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Обновлённое состояние табло
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scoreboard'
  /api/scoreboard/{game_id}/swap-sides:
    post:
      summary: Смена сторон площадок
      description: Меняет home_side / away_side и очищает уведомления о смене
      tags:
        - Scoreboard
      operationId: swapSides
      parameters:
        - name: game_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Обновлённое состояние табло
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scoreboard'
  /api/scoreboard/{game_id}/period:
    post:
      summary: Изменить период/сет
      description: Переключает период (вперёд/назад), обнуляет счёт, фиксирует фолы
      tags:
        - Scoreboard
      operationId: updatePeriod
      parameters:
        - name: game_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - delta
              properties:
                delta:
                  type: integer
                  enum: [1, -1]
                  description: +1 — следующий период, -1 — предыдущий
      responses:
        '200':
          description: Обновлённое состояние табло
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scoreboard'
        '400':
          description: Период вне допустимого диапазона
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/scoreboard/{game_id}/display:
    post:
      summary: Настроить отображение табло
      tags:
        - Scoreboard
      operationId: updateDisplay
      parameters:
        - name: game_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - show
              properties:
                show:
                  type: integer
                  enum: [0, 1, 2, 4, 6, 14]
                  description: Режим отображения (битовая маска элементов)
      responses:
        '200':
          description: Обновлённое состояние табло
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scoreboard'
  /api/scoreboard/{game_id}/label:
    post:
      summary: Обновить кастомную метку табло
      tags:
        - Scoreboard
      operationId: updateLabel
      parameters:
        - name: game_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - custom_label
              properties:
                custom_label:
                  type: string
                  description: Текст метки, отображаемый на табло
      responses:
        '200':
          description: Обновлённое состояние табло
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scoreboard'
  /api/scoreboard/{game_id}/teams:
    put:
      summary: Обновить информацию о командах
      tags:
        - Scoreboard
      operationId: updateTeams
      parameters:
        - name: game_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamsUpdate'
      responses:
        '200':
          description: Обновлённое состояние табло
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scoreboard'
  /api/scoreboard/{game_id}/settings:
    patch:
      summary: Обновить настройки табло
      description: Обновляет `invert_tablo`, `unlimited_score`, `two_wins_mode`
      tags:
        - Scoreboard
      operationId: updateSettings
      parameters:
        - name: game_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsUpdate'
      responses:
        '200':
          description: Обновлённое состояние табло
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scoreboard'
  /api/scoreboard/{game_id}/mode:
    post:
      summary: Переключить режим (пляжный/классический)
      tags:
        - Scoreboard
      operationId: toggleMode
      parameters:
        - name: game_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - beach_mode
              properties:
                beach_mode:
                  type: boolean
                  description: true — пляжный волейбол, false — классический
      responses:
        '200':
          description: Обновлённое состояние табло
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scoreboard'
  /api/scoreboard/{game_id}/reset:
    post:
      summary: Сбросить табло
      description: Обнуляет счёт, сеты, команды (имена и цвета сохраняются если keep_settings=true)
      tags:
        - Scoreboard
      operationId: resetScoreboard
      parameters:
        - name: game_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                keep_settings:
                  type: boolean
                  default: true
                  description: Сохранить имена и цвета команд
      responses:
        '200':
          description: Сброшенное состояние табло
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scoreboard'

  # ===================== MATCHES =====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer-токен, полученный через `POST /api/auth/login`.
        
        **Пример:**
        ```
        curl -X POST /api/auth/login -d '{"identity":"admin","password":"..."}'
        # → { "token": "eyJ..." }
        
        curl /api/auth/me -H "Authorization: Bearer eyJ..."
        ```
  schemas:
...
    Scoreboard:
      type: object
      properties:
        id:
          type: string
          description: game_id
        home_team:
          type: string
          example: Динамо
        away_team:
          type: string
          example: Зенит
        home_color:
          type: string
          example: '#ff0000'
        away_color:
          type: string
          example: '#00ff00'
        tournament_name:
          type: string
          example: НВЛ
        venue:
          type: string
          description: Название зала / площадки
          example: Спорткомплекс «Олимп»
        home_score:
          type: integer
          example: 15
        away_score:
          type: integer
          example: 12
        home_sets:
          type: integer
          example: 1
        away_sets:
          type: integer
          example: 0
        home_fouls:
          type: integer
          example: 1
        away_fouls:
          type: integer
          example: 0
        home_timeouts:
          type: integer
          description: Количество использованных таймаутов домашней команды
          example: 1
        away_timeouts:
          type: integer
          description: Количество использованных таймаутов гостевой команды
          example: 0
        current_period:
          type: integer
          example: 2
        beach_current_set:
          type: integer
          example: 1
        period_count:
          type: integer
          example: 5
        beach_mode:
          type: boolean
          example: false
        beach_match_finished:
          type: boolean
          example: false
        classic_match_finished:
          type: boolean
          example: false
        two_wins_mode:
          type: boolean
          description: Режим «две победы» (классический)
          example: false
        unlimited_score:
          type: boolean
          description: Безлимитный счёт без проверки победы в сете
          example: false
        invert_tablo:
          type: boolean
          example: false
        show:
          type: integer
          example: 1
        custom_label:
          type: string
          example: Табло
        home_side:
          type: string
          enum: [left, right]
          example: left
        away_side:
          type: string
          enum: [left, right]
          example: right
        set_history:
          type: array
          items:
            type: object
            properties:
              home:
                type: integer
              away:
                type: integer
        beach_switch_message:
          type: string
        classic_switch_needed:
          type: boolean
        classic_switch_message:
          type: string
        classic_switch_shown:
          type: boolean
        classic_tiebreak_switch_done:
          type: boolean
        lastEdited:
          type: string
          format: date-time
        created:
          type: string
          format: date-time
    CreateScoreboard:
      type: object
      required:
        - game_id
      properties:
        game_id:
          type: string
          description: Уникальный идентификатор игры
          example: game1
        home_team:
          type: string
          default: Team1
        away_team:
          type: string
          default: Team2
        home_color:
          type: string
          default: '#ff0000'
        away_color:
          type: string
          default: '#00ff00'
        tournament_name:
          type: string
          default: НВЛ
        venue:
          type: string
          description: Название зала / площадки
        custom_label:
          type: string
          default: Табло
        beach_mode:
          type: boolean
          default: false
        two_wins_mode:
          type: boolean
          default: false
        invert_tablo:
          type: boolean
          default: false
        unlimited_score:
          type: boolean
          default: false
    ScoreboardUpdate:
      type: object
      description: |
        Разрешённые поля для обновления:
        home_team, away_team, home_color, away_color, tournament_name, venue,
        custom_label, show, home_score, away_score, home_fouls, away_fouls,
        home_sets, away_sets, current_period, period_count, home_side, away_side,
        beach_mode, beach_current_set, beach_switch_message, beach_match_finished,
        classic_match_finished, classic_tiebreak_switch_done, two_wins_mode,
        invert_tablo, unlimited_score, set_history, home_timeouts, away_timeouts
      properties:
        home_team:
          type: string
        away_team:
          type: string
        home_color:
          type: string
        away_color:
          type: string
        tournament_name:
          type: string
        venue:
          type: string
        custom_label:
          type: string
        show:
          type: integer
          enum: [0, 1, 2, 4, 6, 14]
        home_score:
          type: integer
        away_score:
          type: integer
        home_fouls:
          type: integer
        away_fouls:
          type: integer
        home_sets:
          type: integer
        away_sets:
          type: integer
        current_period:
          type: integer
        period_count:
          type: integer
        home_side:
          type: string
          enum: [left, right]
        away_side:
          type: string
          enum: [left, right]
        beach_mode:
          type: boolean
        beach_current_set:
          type: integer
        beach_switch_message:
          type: string
        beach_match_finished:
          type: boolean
        classic_match_finished:
          type: boolean
        classic_tiebreak_switch_done:
          type: boolean
        two_wins_mode:
          type: boolean
        invert_tablo:
          type: boolean
        unlimited_score:
          type: boolean
        set_history:
          type: array
          items:
            type: object
            properties:
              home:
                type: integer
              away:
                type: integer
        home_timeouts:
          type: integer
        away_timeouts:
          type: integer
    TeamsUpdate:
      type: object
      properties:
        home_team:
          type: string
        home_color:
          type: string
        away_team:
          type: string
        away_color:
          type: string
        tournament_name:
          type: string
        venue:
          type: string
          description: Название зала / площадки
    SettingsUpdate:
      type: object
      properties:
        invert_tablo:
          type: boolean
          description: Инвертировать расположение команд
        unlimited_score:
          type: boolean
          description: Безлимитный счёт
        two_wins_mode:
          type: boolean
          description: Режим «две победы» (классический волейбол, 3 сета максимум)

    # ---------- Matches ----------
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: game_id is required